HELPS-788: Omit unnecessary claims from User Aptitude JWT
This commit adds an optional includeEmail parameter to the User Aptitude
jwt function. When this parameter is set to true, the JWT will include
the current user's email address. When the parameter is unset, or false,
the email address will not be included in the JWT's claims, limiting
potential PII exposure.